修改数量的相关数据类型为double

+++ hikyuu_pywrap/_StockTypeInfo.cpp
@@ -14,19 +14,20 @@ using namespace hku;
 
 void export_StockTypeInfo() {
     class_<StockTypeInfo>("StockTypeInfo", init<>())
-            .def(init<uint32, const string&, price_t, price_t, int, size_t, size_t>())
-            //.def(self_ns::str(self))
-            .def("__str__", &StockTypeInfo::toString)
-            .add_property("type", &StockTypeInfo::type)
-            .add_property("description", make_function(&StockTypeInfo::description, return_value_policy<copy_const_reference>()))
-            .add_property("tick", &StockTypeInfo::tick)
-            .add_property("tickValue", &StockTypeInfo::tickValue)
-            .add_property("unit", &StockTypeInfo::unit)
-            .add_property("precision", &StockTypeInfo::precision)
-            .add_property("minTradeNumber", &StockTypeInfo::minTradeNumber)
-            .add_property("maxTradeNumber", &StockTypeInfo::maxTradeNumber)
+      .def(init<uint32, const string&, price_t, price_t, int, double, double>())
+      //.def(self_ns::str(self))
+      .def("__str__", &StockTypeInfo::toString)
+      .add_property("type", &StockTypeInfo::type)
+      .add_property("description", make_function(&StockTypeInfo::description,
+                                                 return_value_policy<copy_const_reference>()))
+      .add_property("tick", &StockTypeInfo::tick)
+      .add_property("tickValue", &StockTypeInfo::tickValue)
+      .add_property("unit", &StockTypeInfo::unit)
+      .add_property("precision", &StockTypeInfo::precision)
+      .add_property("minTradeNumber", &StockTypeInfo::minTradeNumber)
+      .add_property("maxTradeNumber", &StockTypeInfo::maxTradeNumber)
 #if HKU_PYTHON_SUPPORT_PICKLE
-            .def_pickle(normal_pickle_suite<StockTypeInfo>())
+      .def_pickle(normal_pickle_suite<StockTypeInfo>())
 #endif
-            ;
+      ;
 }
